introduce NRCOptimization.cpp
