+++ src/QMCDrivers/WFOpt/QMCFixedSampleLinearOptimizeBatched.cpp
@@ -31,6 +31,7 @@
 #include "Numerics/MatrixOperators.h"
 #include "EstimatorInputDelegates.h"
 #include "Message/UniformCommunicateError.h"
+#include "Numerics/DeterminantOperators.h"
 #include "LinearMethod.h"
 #include <cassert>
 #ifdef HAVE_LMY_ENGINE
